src/server/requirements.txt
python-dateutil==2.7.3
simplejson==3.16.0
jsonschema==3.2.0
paho-mqtt==1.6.1
Flask==2.1.2
python-dotenv==0.20.0
requests==2.22.0
requests-unixsocket==0.2.0